Started major refactoring work.
Incremented the version number. Note that the code is likely broken at
the moment. I've reorganized the contents of the tealpkg Python package
to break the code into logical pieces, making it easier to find bugs and
figure out what individual parts actually do.

Moving forward, the MVC pattern will be used to separate the
input/output portion of the code from the underlying logic and data.
This way, other types of user interfaces (perhaps even a GUI) can be
added later.

In preparation for rewriting the tealbuild code in Python, the outer
main.py code that lives outside the package now uses a multi-call
architecture. This way, additional executables can be added to the
user-facing side of the application simply by adding symbolic links.
This approach might be revisited in the future.

Within the CLI code, Table and StatusLine have been split out into
separate modules, improving ease of maintenance.