Optimized parse_manifest slightly.
The good news is that I could shave about 0.3s off the execution time
for parse_manifest by making some adjustments to the algorithm. The bad
news is that it's still well over 3s on my workstation when parsing the
Slackware64 base package manifest. Furthermore, it's using over 600 MiB
of RAM just storing data for that single manifest.

It appears that a major redesign of this code is necessary. Some
analysis needs to be done to determine whether or not this work should
be done in Python or moved back to shell+awk.